Commit New Version of Demo. Updated Key mapping/Script Organization
Major scripts for player are combined into 3 scripts:
PlayerAbility, PlayerManager, and PlayerStats

Some values are reused from the old script

Dmg in relation to Health is not accurate (not implemented for working,
but for testing)

Knock Back in PlayerAttack() (Now use Circle to check for enemy game
object in the range and take them in as list)

Other Methods remain intact and working. Should run smoother.