Fix Some Small Part of the scripts, Added a "SafeLight" object with Prefab Avaliable
I am encountering some problems getting enemy dmg to work on player in
the scene i am testing it on (the one label with Kuan). Otherwise the
LightZone is a script that attaches to the player object. In which when
the player walks into the lightzone the condition of the status are
triggers (playerstats) making her invincible while inside (takes no dmg)
also (restore hp) once leaving the zone the player lost the invincible
status and can be damaged again. I can not test this until the dmg to
player problem is resolved though.  Let me know if its the set up or a
scripting problem.